Updated exports of functions
